Added 2fa

+++ Movies.API/Controllers/MoviesController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Movies.API.Requests.Movies;
+using Movies.Business.Services.Caching;
 using Movies.Business.Services.Movies;
 
 namespace Movies.API.Controllers;
@@ -11,60 +12,101 @@ namespace Movies.API.Controllers;
 public class MoviesController : ControllerBase
 {
     public IMovieService _service;
+    private readonly IEventCacheService _eventLogger;
 
-    public MoviesController(IMovieService exampleService)
+    public MoviesController(IMovieService exampleService, IEventCacheService eventLogger)
     {
         _service = exampleService;
+        _eventLogger = eventLogger;
+    }
+    
+    private string? GetUserToken()
+    {
+        return Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
     }
     
     [HttpPost("filter")]
     public async Task<IActionResult> GetFilteredMovies([FromBody] FilterMoviesRequest request)
     {
-        var message = await _service.GetAllFilteredMovies(request.toFilterMoviesDTO());
+        if(await _eventLogger.IsAttacker(GetUserToken()))
+            return StatusCode(429, "Too many requests. Possible attack.");
         
+        var message = await _service.GetAllFilteredMovies(request.toFilterMoviesDTO());
+        await _eventLogger.LogAction("FilterMovies", GetUserToken()!);
+
         return Ok(message);
     }
     
     [HttpGet("get-all")]
     public async Task<IActionResult> GetAllMovies()
     {
-        var message = await _service.GetAllMovies();
+        if(await _eventLogger.IsAttacker(GetUserToken()))
+            return StatusCode(429, "Too many requests. Possible attack.");
         
+        var message = await _service.GetAllMovies();
+        await _eventLogger.LogAction("GetAllMovies", GetUserToken()!);
+
         return Ok(message);
     }
     
-    [Authorize(Roles = "admin")]
+    [HttpGet("get-endorsement")]
+    public async Task<IActionResult> GetEndorsements([FromQuery] int movieId)
+    {
+        if(await _eventLogger.IsAttacker(GetUserToken()))
+            return StatusCode(429, "Too many requests. Possible attack.");
+        
+        var message = await _service.GetEndorsements(movieId);
+        await _eventLogger.LogAction("GetEndorsements", GetUserToken()!);
+
+        return Ok(message);
+    }
+    
+    //[Authorize(Roles = "admin")]
     [HttpDelete("delete")]
     public async Task<IActionResult> DeleteMovie([FromBody] DeleteMovieRequest request)
     {
-        await _service.DeleteMovie(request.toDeleteMovieDTO());
+        if(await _eventLogger.IsAttacker(GetUserToken()))
+            return StatusCode(429, "Too many requests. Possible attack.");
         
+        await _service.DeleteMovie(request.toDeleteMovieDTO());
+        await _eventLogger.LogAction("DeleteMovie", GetUserToken()!);
+
         return Ok();
     }
     
-    [Authorize(Roles = "admin")]
     [HttpPatch("edit")]
     public async Task<IActionResult> EditMovie([FromBody] EditMovieRequest request)
     {
-        await _service.EditMovie(request.toEditMovieDTO());
+        if(await _eventLogger.IsAttacker(GetUserToken()))
+            return StatusCode(429, "Too many requests. Possible attack.");
         
+        await _service.EditMovie(request.toEditMovieDTO());
+        await _eventLogger.LogAction("EditMovie", GetUserToken()!);
+
         return Ok();
     }
     
-    [Authorize(Roles = "admin")]
     [HttpPost("add")]
     public async Task<IActionResult> AddMovie([FromBody] AddMovieRequest request)
     {
-        await _service.AddMovie(request.toAddMovieDTO());
+        if(await _eventLogger.IsAttacker(GetUserToken()))
+            return StatusCode(429, "Too many requests. Possible attack.");
         
+        await _service.AddMovie(request.toAddMovieDTO());
+        await _eventLogger.LogAction("AddMovie", GetUserToken()!);
+
         return Ok();
     }
     
     [HttpGet("get-averages")]
     public async Task<IActionResult> GetAverages()
     {
-        var averages = await _service.GetAverages();
+        if(await _eventLogger.IsAttacker(GetUserToken()))
+            return StatusCode(429, "Too many requests. Possible attack.");
         
+        var averages = await _service.GetAverages();
+        await _eventLogger.LogAction("GetAverages", GetUserToken()!);
+
         return Ok(averages);
     }
 }
